% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignText.R
\name{alignText}
\alias{alignText}
\title{Align strings in two places}
\usage{
alignText(
  leftSide,
  rightSide,
  prependBy = NULL,
  asExpression = FALSE,
  device = NULL,
  ...
)
}
\arguments{
\item{leftSide}{string that should be left aligned}

\item{rightSide}{string that should be left aligned and concatenated with}

\item{prependBy}{a string to prependBy}

\item{asExpression}{if TRUE, then the returned string is suitabled to be parsed as expression.}

\item{device}{a graphics device}

\item{...}{further arguments passed to the device.}
}
\value{
a vector of strings
}
\description{
Align strings in two places
}
