% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspLabelAxes.R
\name{axesLabeller}
\alias{axesLabeller}
\title{Parse numeric to character in a visually appeasing way.}
\usage{
axesLabeller(x, ...)
}
\arguments{
\item{x}{numeric values to parse.}

\item{...}{passed to subfunctions.}
}
\description{
Parse numeric to character in a visually appeasing way.
}
\details{
Useful to avoid weird labels when dealing with very
large or very small numbers. Mainly a wrapper around scales::scientific_format and scales::scientific_format
scales::number_format.
}
