% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPizza.R
\name{drawBFpizza}
\alias{drawBFpizza}
\title{Make a pizza plot}
\usage{
drawBFpizza(
  dat,
  linewidth = 1,
  scaleText = 0.3,
  show.legend = FALSE,
  labels = NULL
)
}
\arguments{
\item{dat}{data.frame with \code{$y}, or vector.}

\item{linewidth}{Scalar, linewidth of the black lines around the pizza.}

\item{scaleText}{Scalar > 0, the size of the text is equal to \code{scaleText * getGraphOption("fontsize")}.}

\item{show.legend}{Logical, should a legend be shown?}

\item{labels}{String vector of length 2, text to be shown above and below plot.}
}
\description{
Make a pizza plot
}
\examples{
JASPgraphs::drawBFpizza(c(1, 4))

JASPgraphs::drawBFpizza(
  dat = data.frame(y = c(1, 4)),
 labels = c("data | H0", "data | H1")
)

}
