% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drawLines}
\alias{drawLines}
\title{Deprecated: use \code{ggplot2::geom_line} instead.}
\usage{
drawLines(
  graph = drawAxis(),
  dat,
  mapping = NULL,
  size = 1.25,
  alpha = 1,
  show.legend = TRUE,
  ...
)
}
\arguments{
\item{graph}{ggplot2 object}

\item{dat}{data frame}

\item{mapping}{mapping from aes}

\item{size}{size}

\item{alpha}{transparancy}

\item{show.legend}{show legend?}

\item{...}{other arguments to geom_line}
}
\description{
Deprecated: use \code{ggplot2::geom_line} instead.
}
