% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drawPoints}
\alias{drawPoints}
\title{Deprecated: use \code{ggplot2::geom_point} instead.}
\usage{
drawPoints(
  graph = drawAxis(),
  dat,
  mapping = NULL,
  size = 1.25,
  shape = 21,
  fill = "gray",
  show.legend = TRUE,
  ...
)
}
\arguments{
\item{graph}{ggplot2 object}

\item{dat}{data frame}

\item{mapping}{mapping from aes}

\item{size}{size}

\item{shape}{shape}

\item{fill}{color for filling}

\item{show.legend}{show legend?}

\item{...}{other arguments to geom_point}
}
\description{
Deprecated: use \code{ggplot2::geom_point} instead.
}
