% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drawSmooth}
\alias{drawSmooth}
\title{Deprecated: use \code{ggplot2::geom_smooth} instead.}
\usage{
drawSmooth(
  graph = NULL,
  dat = NULL,
  mapping = NULL,
  size = 2,
  method = "auto",
  color = "gray",
  show.legend = FALSE,
  se = FALSE,
  alpha = 1,
  ...
)
}
\arguments{
\item{graph}{ggplot2 object}

\item{dat}{data frame}

\item{mapping}{mapping from aes}

\item{size}{size}

\item{method}{statistical method to draw regression line (e.g., lm)}

\item{color}{line color}

\item{show.legend}{show legend?}

\item{se}{show standard errors?}

\item{alpha}{transparancy}

\item{...}{other arguments to geom_smooth}
}
\description{
Deprecated: use \code{ggplot2::geom_smooth} instead.
}
