% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customGeoms.R
\name{geom_point}
\alias{geom_point}
\alias{geom_line}
\title{Custom geoms}
\usage{
geom_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_line(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{see details}

\item{data}{see details}

\item{stat}{see details}

\item{position}{see details}

\item{...}{see details}

\item{na.rm}{see details}

\item{show.legend}{see details}

\item{inherit.aes}{see details}
}
\description{
Custom geoms
}
\details{
These functions are virtually identical to \code{\link[ggplot2]{geom_point}} and \code{\link[ggplot2]{geom_line}}
except that their underlying Geoms have different default values.
}
