% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatability.R
\name{getAxisBreaks}
\alias{getAxisBreaks}
\title{Get the axis breaks of a ggplot2 object}
\usage{
getAxisBreaks(x)
}
\arguments{
\item{x}{a ggplot object}
}
\description{
Get the axis breaks of a ggplot2 object
}
