% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPriorAndPosterior.R
\name{getEmptyTheme}
\alias{getEmptyTheme}
\title{Return an empty theme}
\usage{
getEmptyTheme()
}
\description{
Return an empty theme
}
\details{
a modification of \code{\link[ggplot2]{theme_void}}
}
