% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrettyAxisBreaks.R
\name{getPrettyAxisBreaks}
\alias{getPrettyAxisBreaks}
\title{Compute axis breaks}
\usage{
getPrettyAxisBreaks(x, ...)
}
\arguments{
\item{x}{the object to compute axis breaks for}

\item{...}{if x is numeric, this is passed to pretty}
}
\description{
Compute axis breaks
}
\details{
this is just a wrapper for pretty with slightly different defaults.
}
