% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphOptions.R
\name{.graphOptions}
\alias{.graphOptions}
\alias{getGraphOption}
\alias{setGraphOption}
\alias{graphOptions}
\title{Global graph options}
\usage{
.graphOptions

getGraphOption(name)

setGraphOption(name, value)

graphOptions(...)
}
\arguments{
\item{name}{character string of the value to get or set.}

\item{value}{the value to change x into.}

\item{...}{modify options using name = value.}
}
\description{
Global graph options
}
