% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPriorAndPosterior.R
\name{hypothesis2BFtxt}
\alias{hypothesis2BFtxt}
\title{Get text for Bayes factor pie chart given hypothesis}
\usage{
hypothesis2BFtxt(hypothesis = c("equal", "smaller", "greater"))
}
\arguments{
\item{hypothesis}{hypothesis}
}
\description{
Get text for Bayes factor pie chart given hypothesis
}
