% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRecipe.R
\name{isJaspPlotRecipe}
\alias{isJaspPlotRecipe}
\title{Test whether an object is a plot recipe}
\usage{
isJaspPlotRecipe(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{jaspPlotRecipe}, otherwise
\code{FALSE}.
}
\description{
Test whether an object is a plot recipe
}
\seealso{
\code{\link{createJaspPlotRecipe}}
}
