% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspBivariate.R
\name{jaspBivariate}
\alias{jaspBivariate}
\title{Bivariate plots with optional confidence and prediction intervals.}
\usage{
jaspBivariate(
  x,
  y,
  group = NULL,
  xName,
  yName,
  groupName,
  type = c("point", "hex", "bin", "contour", "density", "none"),
  args = list(),
  smooth = c("none", "lm", "glm", "gam", "loess"),
  smoothCi = FALSE,
  smoothCiLevel = 0.95,
  smoothArgs = list(),
  predict = c("none", "lm", "ellipse"),
  predictLevel = 0.95,
  predictArgs = .predictArgs(),
  xBreaks = NULL,
  yBreaks = NULL,
  legendPosition = "none"
)
}
\arguments{
\item{x}{Numeric vector of values on the x-axis.}

\item{y}{Numeric vector of values on the y-axis.}

\item{group}{Optional grouping variable.}

\item{xName}{Character; x-axis label. If left empty, the name of the \code{x} object is displayed. To remove the axis label, use \code{NULL}.}

\item{yName}{Character; y-axis label. If left empty, the name of the \code{y} object is displayed. To remove the axis label, use \code{NULL}.}

\item{groupName}{Character; label of the grouping variable displayed as a legend title. If left empty, the name of the \code{group} object is displayed.}

\item{type}{Character; How should the distribution of the data be displayed:
\describe{
\item{"point"}{Using \link{geom_point}.}
\item{"hex"}{Using \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}
\item{"bin"}{Using \link[ggplot2:geom_bin2d]{ggplot2::geom_bin2d}.}
\item{"contour"}{Using \link[ggplot2:geom_density2d]{ggplot2::geom_density2d}.}
\item{"density"}{Using \link[ggplot2:geom_density2d_filled]{ggplot2::geom_density2d_filled}.}
}}

\item{args}{A list of additional arguments passed to the geom function determined by \code{type} argument.}

\item{smooth}{Character; passed as \code{method} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth},
unless \code{smooth == "none"}, in which case the layer is not plotted.}

\item{smoothCi}{Logical; Should confidence interval around the smooth line be plotted?
Passed as \code{se} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{smoothCiLevel}{Numeric; Confidence level of the confidence interval around the smooth line.
Passed as \code{level} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{smoothArgs}{A list of additional arguments passed to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{predict}{Character; Method for drawing the prediction interval:
\describe{
\item{"none"}{Prediction interval is not displayed.}
\item{"lm"}{Prediction interval is plotted, the confidence bands are calculated using \link[stats:predict.lm]{stats::predict.lm}.}
\item{"ellipse"}{Prediction ellipse is plotted using \link[ggplot2:stat_ellipse]{ggplot2::stat_ellipse}.}
}}

\item{predictLevel}{Numeric; Confidence level of the prediction interval.}

\item{predictArgs}{A list of additional arguments passed to the function that draws the prediction interval.}

\item{xBreaks}{Optional numeric vector that specifies the breaks along the x-axis.}

\item{yBreaks}{Optional numeric vector that specifies the breaks along the y-axis.}

\item{legendPosition}{Character; passed as \code{legend.position} to \link{themeJaspRaw}.}
}
\description{
This plot consists of three layers:
\enumerate{
\item The bivariate distribution.
\item Smooth line through the data displayed using \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.
\item Prediction interval of y given x using \link[stats:predict.lm]{stats::predict.lm}(assuming linear relationship), or prediction ellipse assuming bivariate normal distribution.
}
}
