% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JASPgraphsPlot.R
\name{JASPgraphsPlotMethods}
\alias{JASPgraphsPlotMethods}
\alias{[[JASPgraphsPlot}
\alias{[[<-JASPgraphsPlot}
\alias{is.JASPgraphsPlot}
\alias{length.JASPgraphsPlot}
\alias{names.JASPgraphsPlot}
\title{Methods for interacting with a JASPgraphsPlot}
\usage{
\method{[[}{JASPgraphsPlot}(x, field)

\method{[[}{JASPgraphsPlot}(x, field) <- value

is.JASPgraphsPlot(x)

\method{length}{JASPgraphsPlot}(x)

\method{names}{JASPgraphsPlot}(x)
}
\arguments{
\item{x}{an object of class JASPgraphsPlot}

\item{field}{the name or index of a subplot}

\item{value}{the value that should be assigned}
}
\description{
These methods are mainly convenience functions that ensure things like seq_along work.
}
