% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspHistogram.R
\name{jaspHistogram}
\alias{jaspHistogram}
\title{jaspHistogram}
\usage{
jaspHistogram(
  x,
  xName,
  rugs = FALSE,
  displayDensity = FALSE,
  binWidthType = c("doane", "fd", "scott", "sturges", "manual"),
  numberOfBins = NA,
  densityLineWidth = 1
)
}
\arguments{
\item{x, }{numeric, the data to show a histogram for}

\item{xName, }{string, the title on the x-axis}

\item{rugs, }{logical, should rugs be shown on the x-axis?}

\item{displayDensity, }{logical, should a density be superimposed on the plot?}

\item{binWidthType, }{string, type of bindwidth, matches the string values that can be passed to \code{\link[graphics]{hist}}.}

\item{numberOfBins, }{if binWidthType is set to "manual", this value determines the number of bins.}

\item{densityLineWidth, }{the line width of the superimposed density.}
}
\description{
jaspHistogram
}
\examples{
set.seed(1234)
x <- rnorm(1e3) + rnorm(1e3, 1, .1) - rgamma(1e3, 10, 5)
jaspHistogram(x, "random data")
jaspHistogram(x, "random data", displayDensity = TRUE)
jaspHistogram(x, "random data", displayDensity = TRUE, rugs = TRUE)
jaspHistogram(x, "random data", rugs = TRUE, binWidthType = "sturges")
}
