% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{jaspScatter}
\alias{jaspScatter}
\title{Scatter plot}
\usage{
jaspScatter(
  x,
  y,
  xName = NULL,
  yName = NULL,
  type = c("point", "hex", "auto"),
  bins = 30,
  binwidth = NULL,
  palette = "gray",
  fill = "gray",
  alpha = 1,
  smooth = c("none", "lm", "glm", "gam", "loess"),
  smoothCi = FALSE,
  smoothCiLevel = 0.95,
  smoothColor = adjustcolor("darkred", alpha = 0.5),
  predict = c("none", "lm", "ellipse"),
  predictCiLevel = 0.95,
  predictColor = adjustcolor("steelblue", alpha = 0.5),
  suppressAxesLabels = FALSE
)
}
\arguments{
\item{x}{Numeric vector of values on the x-axis.}

\item{y}{Numeric vector of values on the y-axis.}

\item{xName}{Character; x-axis label.}

\item{yName}{Character; y-axis label.}

\item{type}{Character; How should the distribution of the data be displayed:
\describe{
\item{"point"}{Using \link{geom_point}.}
\item{"hex"}{Using \link[ggplot2:geom_hex]{ggplot2::geom_hex}}
\item{"auto"}{If the number of data points is smaller than 1000, using \link{geom_point}, otherwise using \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}
}}

\item{bins, binwidth}{Arguments passed to \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}

\item{palette}{Argument passed to \link{JASPcolors}. Palette to use for drawing \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}

\item{fill}{Argument passed to \link{geom_point}.}

\item{alpha}{Argument passed to \link{geom_point} or \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}

\item{smooth}{Character; passed as \code{method} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth},
unless \code{smooth == "none"}, in which case the layer si not plotted.}

\item{smoothCi}{Logical; Should confidence interval around the smooth line be plotted?
Passed as \code{se} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{smoothCiLevel}{Numeric; Confidence level of the confidence interval around the smooth line.
Passed as \code{level} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{smoothColor}{Color of the smooth line.}

\item{predict}{Character; Method for drawing the prediction interval:
\describe{
\item{"none"}{Prediction interval is not displayed.}
\item{"lm"}{Prediction interval is plotted, the confidence bands are calculated using \link[stats:predict.lm]{stats::predict.lm}.}
\item{"ellipse"}{Prediction ellipse is plotted, assuming bi-variate normal model.}
}}

\item{predictCiLevel}{Numeric; Confidence level of the prediction interval.}

\item{predictColor}{Color of the prediction interval.}

\item{suppressAxesLabels}{Logical; should axis labels be suppressed.}
}
\description{
This plot consists of three layers:
\enumerate{
\item The distribution of the data displayed as \link{geom_point} or as a \link[ggplot2:geom_hex]{ggplot2::geom_hex}.
\item Smooth line through the data displayed using \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.
\item Prediction interval of y given x using \link[stats:predict.lm]{stats::predict.lm}(assuming linear relationship), or prediction ellipse assuming bivariate normal distribution.
}
}
