% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEditingOptions.R
\name{plotEditingOptions}
\alias{plotEditingOptions}
\title{Get the editable options for a graph}
\usage{
plotEditingOptions(graph, asJSON = FALSE)
}
\arguments{
\item{graph}{a plot object}

\item{asJSON}{should the list be converted to JSON?}
}
\description{
Get the editable options for a graph
}
