% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPriorAndPosterior.R
\name{PlotPriorAndPosterior}
\alias{PlotPriorAndPosterior}
\title{Create a prior-posterior plot.}
\usage{
PlotPriorAndPosterior(dfLines, dfPoints = NULL, BF01 = NULL,
  CRI = NULL, median = NULL, xName = NULL, yName = "Density",
  drawPizzaTxt = !is.null(BF01), drawCRItxt = !is.null(CRI),
  bfSubscripts = "BF[0][1]", pizzaTxt = pizzaTxtFromBF(bfSubscripts),
  bty = list(type = "n", ldwX = 0.5, lwdY = 0.5), ...)
}
\arguments{
\item{dfLines}{A dataframe with \code{$x}, \code{$y}, and optionally \code{$g}.}

\item{dfPoints}{A dataframe with \code{$x}, \code{$y}, and optionally \code{$g}.}

\item{BF01}{Numeric, with value of Bayes factor.}

\item{CRI}{Numeric of length 2, Credible interval of posterior.}

\item{median}{Numeric, median of posterior.}

\item{xName}{String or expression, displayed on the x-axis.}

\item{yName}{String or expression, displayed on the y-axis.}

\item{drawPizzaTxt}{Logical, should there be text above and below the pizza plot?}

\item{drawCRItxt}{Logical, should the credible interval be displayed in text?}

\item{bfSubscripts}{String, to be parsed as expression. Indicates what BF type to display.}

\item{pizzaTxt}{String vector of length 2, text to be drawn above and below pizza plot.}

\item{bty}{List of three elements. Type specifies the box type, ldwX the width of the x-axis, lwdY the width of the y-axis.}

\item{...}{Unused.}
}
\description{
Create a prior-posterior plot.
}
