% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEnvironments.R
\name{removeEnvironments}
\alias{removeEnvironments}
\title{removeEnvironments - remove environments from ggplot2 objects}
\usage{
removeEnvironments(x, debugInfo = 0, ...)
}
\arguments{
\item{x}{a plot object. Currently, only ggplot2 objects and jaspGraphsPlot are supported, other objects are returned unchanged}

\item{debugInfo}{Should debugInfo be printed? Possible values are 0, no info is shown; 1, some info is shown; or 2, a lot of info is shown.}

\item{...}{internal.}
}
\value{
an object of the same class as the input
}
\description{
removeEnvironments - remove environments from ggplot2 objects
}
\details{
This function may breaks ggplot2 objects that rely on looking up values in environments.
see also https://github.com/tidyverse/ggplot2/issues/3994 and https://github.com/tidyverse/ggplot2/issues/4056
}
\examples{
# adapted from https://github.com/tidyverse/ggplot2/issues/4056
\dontrun{
demo <- function(n = 1e5) {
  big_object <- rnorm(n)
  dt <- data.frame(x = 1:10)
  p <- ggplot2::ggplot(dt, ggplot2::aes(x, y)) +
    ggplot2::geom_point() +
    ggplot2::geom_point(ggplot2::aes(x = x+1, y = y+1))
}

sizeOnDisk <- function(obj) {
  f <- tempfile()
  saveRDS(obj)
  res <- file.size(f)
  file.remove(f)
  res
}

obj <- demo()
# object.size lies, but lobstr::obj_size does not
object.size(obj)
object.size(obj$plot_env$big_object)

if (require("lobstr"))
  lobstr::obj_size(obj)

sizeOnDisk(obj)

# note that to an extent, removeEnvironments modifies in place
obj <- removeEnvironments(obj)

sizeOnDisk(obj)


}
}
