% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeJasp.R
\name{themeJaspRaw}
\alias{themeJaspRaw}
\title{JASP theme for ggplot2 objects}
\usage{
themeJaspRaw(
  legend.position = "none",
  legend.cex = 1,
  axis.title.cex = 1,
  family = getGraphOption("family"),
  axisTickLength = getGraphOption("axisTickLength"),
  axisTickWidth = getGraphOption("axisTickWidth"),
  fontsize = getGraphOption("fontsize"),
  legend.justification = "top",
  legend.title = element_text(family = family, size = fontsize, hjust = 0.5),
  Xvjust = NULL,
  Yvjust = NULL
)
}
\arguments{
\item{legend.position}{where should the legend be?}

\item{legend.cex}{magnification of the font size for the legend.}

\item{axis.title.cex}{scalar magnification for the title of the axes.}

\item{family}{global font familiy.}

\item{axisTickLength}{length of axis ticks.}

\item{axisTickWidth}{width of axis ticks.}

\item{fontsize}{global font size.}

\item{legend.justification}{see theme}

\item{legend.title}{should legend have a title?}

\item{Xvjust}{adjustment of x-axis tick labels}

\item{Yvjust}{adjustment of y-axis tick labels}
}
\description{
JASP theme for ggplot2 objects
}
