% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePubr.R
\name{themePubrRaw}
\alias{themePubrRaw}
\title{Pubr theme for ggplot2 objects based on \code{ggpubr::theme_pubr}}
\usage{
themePubrRaw(
  base_size = getGraphOption("fontsize"),
  base_family = getGraphOption("family"),
  border = FALSE,
  margin = TRUE,
  legend = getGraphOption("legend.position"),
  x.text.angle = 0,
  axisTickLength = getGraphOption("axisTickLength"),
  axisTickWidth = getGraphOption("axisTickWidth")
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{border}{logical value. Default is FALSE. If TRUE, add panel border.}

\item{margin}{logical value. Default is TRUE. If FALSE, reduce plot margin.}

\item{legend}{character specifying legend position. Allowed values are one of
c("top", "bottom", "left", "right", "none"). Default is "top" side position.
to remove the legend use legend = "none". Legend position can be also
specified using a numeric vector c(x, y).  In this case it is possible to
position the legend inside the plotting area. x and y are the coordinates of
the legend box. Their values should be between 0 and 1. c(0,0) corresponds
to the "bottom left" and c(1,1) corresponds to the "top right" position. For
instance use legend = c(0.8, 0.2).}

\item{x.text.angle}{Rotation angle of x axis tick labels. Default value is 0.
Use 90 for vertical text.}

\item{axisTickLength}{length of axis ticks.}

\item{axisTickWidth}{width of axis ticks.}
}
\description{
Set a Pubr theme and additional options like other themes,
\code{graph + themePubrRaw()}.

This function is a modification of
 \code{\link[ggpubr:theme_pubr]{ggpubr::theme_pubr}} that create a publication ready
 theme.
}
