% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPalettes.R
\name{JASPcolors}
\alias{JASPcolors}
\alias{scale_JASPcolor_continuous}
\alias{scale_JASPfill_continuous}
\alias{scale_JASPcolor_discrete}
\alias{scale_JASPfill_discrete}
\title{JASP color palettes}
\usage{
JASPcolors(n, palette = getGraphOption("palette")[["palette"]],
  asFunction = FALSE)


  scale_JASPcolor_continuous(palette = getGraphOption("palette")[["Function"]],
  ...)


  scale_JASPfill_continuous(palette = getGraphOption("palette")[["Function"]],
  ...)


  scale_JASPcolor_discrete(palette = getGraphOption("palette")[["Function"]],
  ...)


  scale_JASPfill_discrete(palette = getGraphOption("palette")[["Function"]],
  ...)
}
\arguments{
\item{n}{Number of colors desired.}

\item{palette}{Palette to choose from.}

\item{asFunction}{Should a function be returned or the raw colors? If a function is returned, it either takes a single integer or a vector in 0, 1 as input.}

\item{...}{Further arguments for \code{\link{ggplot2::scale_color_manual}} or \code{\link{ggplot2::scale_fill_manual}}.}
}
\value{
Either a character vector of colors or a function.
}
\description{
JASP color palettes
}
\details{
For ggplot2s, the convenience functions \code{scale_JASPcolor_\*} and \code{scale_JASPfill_\*} exist.
}
\examples{
library(ggplot2)

# use colors directly
colors <- JASPcolors(10, "colorblind")
plot(rnorm(10), col = colors, pch = 16, cex = 2)

colFun <- JASPcolors(palette = "colorblind", asFunction = TRUE)
plot(rnorm(10), col = colFun(10), pch = 16, cex = 2)

# use colorscales
e0 <- 1:10
df <- data.frame(x = e0, y = e0, z = e0, g = factor(e0))
ggplot(df, aes(x = x, y = y, color = z, fill = z)) +
  geom_point() +
  scale_JASPcolor_continuous() +
  scale_JASPfill_continuous()

ggplot(df, aes(x = x, y = y, color = g, fill = g)) +
  geom_point() +
  scale_JASPcolor_discrete() +
  scale_JASPfill_discrete()
}
