% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGgplotToPlotly.R
\name{convertGgplotToPlotly}
\alias{convertGgplotToPlotly}
\title{convert a ggplot object to a plotly object and store the json in the image list}
\usage{
convertGgplotToPlotly(ggplotObj, returnJSON = TRUE)
}
\arguments{
\item{ggplotObj}{a ggplot object}

\item{returnJSON}{Should a plotly object be returned (FALSE) or a JSON object for JASP (TRUE)?}
}
\description{
convert a ggplot object to a plotly object and store the json in the image list
}
