% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rangeframe.R
\docType{data}
\name{geom_rangeframe}
\alias{geom_rangeframe}
\alias{GeomRangeFrame}
\title{Add base R's bty = 'n' to ggplot2 objects.}
\usage{
geom_rangeframe(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  sides = "bl",
  panelInfo = list(t = "x.sec.major", r = "y.sec.major", b = "x.major", l = "y.major"),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{sides}{A string that controls which sides of the plot the frames appear on.
It can be set to a string containing any of \code{'trbl'}, for top, right,
bottom, and left. By default, only the bottom and left axes lines are drawn.
Note that this is checked at drawing time, so 'b' always means bottom even when using \code{\link[ggplot2]{coord_flip}}.}

\item{panelInfo}{A list that specifies what information is drawn from what component of panel_scales.
Usually, \code{x.major} corresponds to the bottom axis. However, if a scale is used to move e.g.,
the x-axis to be above the plot then this needs to be adjusted to \code{x.sec.major_source}. By default,
this argument assumes that axis lines above and right of a plot use \code{\link[ggplot2]{sec_axis}} and are
therefore draw information from \code{*.sec.major_source}. You can partially set this list; if e.g.,
\code{"t"} is missing it will be filled with its default value.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}
}
\description{
Axis lines which extend to the maximum and minimum of the axis breaks. The implementation and documentation is largely adapted from \code{\link[ggthemes]{geom_rangeframe}}.
}
\section{Aesthetics}{

\itemize{
\item colour
\item size
\item linetype
\item alpha
}
}

\examples{
library(ggplot2)
library(jaspGraphs)

# an example plot
graph <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point()

# add bty = 'n' for x and y-axes.
graph +
 geom_rangeframe() +
 themeJaspRaw()

# automatically done by themeJasp
themeJasp(graph)

# only draw the x-axis line
graph +
  geom_rangeframe(sides = "b") +
  themeJaspRaw()

# coord_flip does not interact with sides
graph + 
  coord_flip() +
  geom_rangeframe(sides = "b") + 
  themeJaspRaw()

# add axis lines right and above of plot:
graph + 
  scale_x_continuous(sec.axis = sec_axis(~ log10(.))) + 
  scale_y_continuous(sec.axis = sec_axis(~ log10(.))) + 
  geom_rangeframe(sides = "trbl") + 
  themeJaspRaw()

# set x-axis to above the plot and draw a line there
graph + 
  scale_x_continuous(position = "top") + 
  geom_rangeframe(sides = "tl", panelInfo = list(t = "x.major")) + 
  themeJaspRaw()
}
\references{
Tufte, Edward R. (2001) The Visual Display of Quantitative Information, Chapter 6.

Jeffrey B. Arnold (2019). ggthemes: Extra Themes, Scales and Geoms for 'ggplot2'. R package version 4.2.0. https://CRAN.R-project.org/package=ggthemes
}
\keyword{datasets}
