% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphOptions.R
\docType{data}
\name{.graphOptions}
\alias{.graphOptions}
\alias{getGraphOption}
\alias{setGraphOption}
\alias{graphOptions}
\alias{setJASPFont}
\title{Global graph options}
\format{
An object of class \code{environment} of length 13.
}
\usage{
.graphOptions

getGraphOption(name)

setGraphOption(name, value)

graphOptions(...)

setJASPFont(path)
}
\arguments{
\item{name}{characyer string of the value to get or set.}

\item{value}{the value to change x into.}

\item{...}{modify options using name = value.}

\item{path}{path to use for JASP_FONT}
}
\description{
Global graph options
}
\keyword{datasets}
