% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspBivariate.R
\name{jaspBivariateWithMargins}
\alias{jaspBivariateWithMargins}
\title{Bivariate plots with marginal distributions along the axes.}
\usage{
jaspBivariateWithMargins(
  x,
  y,
  group = NULL,
  xName,
  yName,
  groupName,
  margins = c(1/6, 5/6),
  xMarginalArgs = .marginalArgs(),
  yMarginalArgs = .marginalArgs(),
  topRightPlotFunction = NULL,
  topRightPlotArgs = list(),
  legendPosition = "topRight",
  ...
)
}
\arguments{
\item{x}{Numeric vector of values on the x-axis.}

\item{y}{Numeric vector of values on the y-axis.}

\item{group}{Optional grouping variable.}

\item{xName}{Character; x-axis label. If left empty, the name of the \code{x} object is displayed. To remove the axis label, use \code{NULL}.}

\item{yName}{Character; y-axis label. If left empty, the name of the \code{y} object is displayed. To remove the axis label, use \code{NULL}.}

\item{groupName}{Character; label of the grouping variable displayed as a legend title. If left empty, the name of the \code{group} object is displayed.}

\item{margins}{Numeric vector; The proportions of the subplots relative to each other.}

\item{xMarginalArgs}{List, options for the marginal plot above. Defaults to the default values of \link{jaspMarginal}.}

\item{yMarginalArgs}{List, options for the marginal plot to the right. Defaults to the default values of \link{jaspMarginal}.}

\item{topRightPlotFunction}{An optional function that can be used to plotting something in the top-right panel. If \code{NULL} (default), an empty area is plotted.}

\item{topRightPlotArgs}{An optional list of options passed to \code{topRightPlotFunction}.}

\item{legendPosition}{Either "topRight" or any values that is accepted by \code{\link[ggplot2]{theme}} for \code{legend.position}. If set to "topRight" then \code{topRightPlotFunction} cannot be used.}

\item{...}{Additional options passed to \link{jaspBivariate}.}
}
\description{
This plot consists of four elements:
\enumerate{
\item The bivariate plot of \code{x} and \code{y} in the bottom-left panel displayed by \link{jaspBivariate}.
\item Marginal distributions along the diagonal displayed by \link{jaspHistogram}. The plot on the bottom-right has transposed axes.
\item (Optional) custom plot on the top-right panel. See details.
}
}
