% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JASPgraphsPlot.R
\name{jaspGraphsPlotMethods}
\alias{jaspGraphsPlotMethods}
\alias{[[.jaspGraphsPlot}
\alias{[[<-.jaspGraphsPlot}
\alias{is.jaspGraphsPlot}
\alias{length.jaspGraphsPlot}
\alias{names.jaspGraphsPlot}
\title{Methods for interacting with a jaspGraphsPlot}
\usage{
\method{[[}{jaspGraphsPlot}(x, field)

\method{[[}{jaspGraphsPlot}(x, field) <- value

is.jaspGraphsPlot(x)

\method{length}{jaspGraphsPlot}(x)

\method{names}{jaspGraphsPlot}(x)
}
\arguments{
\item{x}{an object of class jaspGraphsPlot}

\item{field}{the name or index of a subplot}

\item{value}{the value that should be assigned}
}
\description{
These methods are mainly convenience functions that ensure things like seq_along work.
}
