% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspMarginal.R
\name{jaspMarginal}
\alias{jaspMarginal}
\alias{.marginalArgs}
\alias{.histogramArgs}
\alias{.rugArgs}
\alias{.densityArgs}
\title{Histograms and Density plots for JASP}
\usage{
jaspMarginal(
  x,
  group = NULL,
  xName,
  groupName,
  yName,
  type = c("auto", "count", "density"),
  breaks = "sturges",
  xAxis = NULL,
  histogram = TRUE,
  histogramArgs = .histogramArgs(),
  rug = FALSE,
  rugArgs = .rugArgs(),
  density = FALSE,
  densityArgs = .densityArgs(),
  densityOverlay = FALSE,
  densityOverlayArgs = .densityArgs(linewidth = 1),
  axisLabels = c("auto", "both", "x", "y", "none"),
  sides = "bl"
)

.marginalArgs(
  x,
  group = NULL,
  xName,
  groupName,
  yName,
  type = c("auto", "count", "density"),
  breaks = "sturges",
  xAxis = NULL,
  histogram = TRUE,
  histogramArgs = .histogramArgs(),
  rug = FALSE,
  rugArgs = .rugArgs(),
  density = FALSE,
  densityArgs = .densityArgs(),
  densityOverlay = FALSE,
  densityOverlayArgs = .densityArgs(linewidth = 1),
  axisLabels = c("auto", "both", "x", "y", "none"),
  sides = "bl"
)

.histogramArgs(linewidth = 0.7, position = ggplot2::position_dodge(), ...)

.rugArgs(...)

.densityArgs(linewidth = 0.7, alpha = 0.5, ...)
}
\arguments{
\item{x, }{numeric, the data to show the plot for.}

\item{group, }{factor, show \code{histogram}, \code{density}, and \code{rug} split by groups?}

\item{xName, }{string, the title on the x-axis. Use \code{NULL} to hide the axis title. If \code{base::missing}, the value is inferred from the object name passed to \code{x}.}

\item{groupName, }{character, legend name of the grouping variable. Use \code{NULL} to hide the legend title. If \code{base::missing}, the value is inferred from the object name passed to \code{group}.}

\item{yName, }{string, the title on the y-axis. Use \code{NULL} to hide the axis title. If \code{base::missing}, "Density" or "Count" is used depending on the value of \code{type}.}

\item{type, }{string, should count or density be displayed on the y-axis? If \code{"auto"}, \code{"density"} is used if \code{density} or \code{densityOverlay} is used, otherwise \code{"count"} is used. \code{"count"} preserves marginal densities if split by group, \code{"density"} re-normalizes each subgroup.}

\item{breaks, }{see \code{breaks} from \code{graphics::hist}. Additionally allows \code{"doane"} method.}

\item{xAxis, }{list, optional shared axis of \code{x} with elements \code{binBreaks} (histogram bin breaks),
\code{breaks} (axis breaks), and \code{limits} (axis limits). If \code{binBreaks} is \code{NULL}, the bins are computed from \code{breaks}.
Used by \link{jaspMatrixPlot} and \link{jaspBivariateWithMargins} so that all panels showing the same variable share the x-axis.}

\item{histogram, }{logical, should a histogram be shown?}

\item{histogramArgs, }{list, additional arguments passed to \code{\link[ggplot2]{geom_histogram}}. Use \code{.histogramArgs} to set the options.}

\item{rug, }{logical, should rugs be shown on the x-axis?}

\item{rugArgs, }{list, additional arguments passed to \code{\link[ggplot2]{geom_rug}}. Use \code{.rugArgs} to set the options.}

\item{density, }{logical, should a density be superimposed on the plot?}

\item{densityArgs, }{logical, additional arguments passed to \code{\link[ggplot2]{geom_density}}. Use \code{.densityArgs} to set the options.}

\item{densityOverlay, }{logical, should a density overlay be superimposed on the plot?}

\item{densityOverlayArgs, }{logical, additional arguments passed to \code{\link[ggplot2]{geom_density}}. Use \code{.densityArgs} to set the options.}

\item{axisLabels, }{string, which axes should have labels displayed? If \code{"auto"}, \code{"x"} is used if \code{type == "density"}, otherwise \code{"both"} is used.}

\item{sides, }{string passed to \link{geom_rangeframe}.}

\item{linewidth}{see \code{?ggplot2::aes_linetype_size_shape}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{alpha}{color transparency}
}
\value{
\code{jaspMarginal} returns the ggplot object.
\code{.marginalArgs}, \code{histogramArgs}, \code{.rugArgs}, and \code{.densityArgs} return a list passed as arguments into their respective ggplot geoms.
}
\description{
A plot histogram with four optional components.
\describe{
\item{\code{histogram}}{Histogram which can be tweaked with \code{breaks} and \code{histogramArgs} arguments.}
\item{\code{density}}{Density line(s) which can be tweaked with \code{densityArgs}.}
\item{\code{densityOverlay}}{Density line which can be tweaked with \code{densityOverlayArgs}. Only one line is shown for the full data regardless of whether \code{group} is used.}
\item{\code{rug}}{Rugs underneath the figure which can be tweaked with \code{rugArgs}.}
}

Each of these components can be enabled (or disabled) individually.

\code{.marginalArgs}, \code{histogramArgs}, \code{.rugArgs}, and \code{.densityArgs} are helper functions for specifying the list of tweaking options for the individual components without overriding other default values.
}
\details{
Colors are taken from \code{graphOptions("palette")}.
}
\examples{
set.seed(1)
x <- rnorm(250)
jaspMarginal(x)
jaspMarginal(x, breaks = 5)
jaspMarginal(x, breaks = 'doane')

jaspMarginal(x, density = TRUE, histogram = FALSE, rug = TRUE)


group <- as.factor(sample(letters[1:2], 500, TRUE, prob = c(0.3, 0.7)))
x     <- rnorm(500, mean = c(a = 0, b = 1)[group])

jaspMarginal(x, group, density = TRUE, rug = TRUE) # does not preserve marginal proportions
jaspMarginal(x, group, density = TRUE, rug = TRUE, type = "count") # preserves marginal proportions

# stacked groups
jaspMarginal(
  x, group, density = TRUE, type = "count",
  histogramArgs = .histogramArgs(position = ggplot2::position_stack()),
  densityArgs = .densityArgs(position = ggplot2::position_stack())
  )

jaspMarginal(x, group, densityOverlay = TRUE, type = "count", histogramArgs = .histogramArgs(position = ggplot2::position_stack()))
}
