% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspMarginal.R
\name{jaspMarginal}
\alias{jaspMarginal}
\alias{.marginalArgs}
\alias{.histogramArgs}
\alias{.rugArgs}
\alias{.densityArgs}
\title{Histograms and Density plots for JASP}
\usage{
jaspMarginal(
  x,
  group = NULL,
  xName,
  groupName,
  yName,
  type = c("auto", "count", "density"),
  breaks = "sturges",
  histogram = TRUE,
  histogramArgs = .histogramArgs(),
  rug = FALSE,
  rugArgs = .rugArgs(),
  density = FALSE,
  densityArgs = .densityArgs(),
  densityOverlay = FALSE,
  densityOverlayArgs = .densityArgs(linewidth = 1),
  axisLabels = c("auto", "both", "x", "y", "none")
)

.marginalArgs(
  x,
  group = NULL,
  xName,
  groupName,
  yName,
  type = c("auto", "count", "density"),
  breaks = "sturges",
  histogram = TRUE,
  histogramArgs = .histogramArgs(),
  rug = FALSE,
  rugArgs = .rugArgs(),
  density = FALSE,
  densityArgs = .densityArgs(),
  densityOverlay = FALSE,
  densityOverlayArgs = .densityArgs(linewidth = 1),
  axisLabels = c("auto", "both", "x", "y", "none")
)

.histogramArgs(
  color = "black",
  size = 0.7,
  position = ggplot2::position_dodge(),
  ...
)

.rugArgs(...)

.densityArgs(color = "black", linewidth = 0.7, alpha = 0.5, ...)
}
\arguments{
\item{x, }{numeric, the data to show the plot for.}

\item{group, }{factor, show \code{histogram}, \code{density}, and \code{rug} split by groups?}

\item{xName, }{string, the title on the x-axis. Use \code{NULL} to hide the axis title. If \code{base::missing}, the value is inferred from the object name passed to \code{x}.}

\item{groupName, }{character, legend name of the grouping variable. Use \code{NULL} to hide the legend title. If \code{base::missing}, the value is inferred from the object name passed to \code{group}.}

\item{yName, }{string, the title on the y-axis. Use \code{NULL} to hide the axis title. If \code{base::missing}, "Density" or "Count" is used depending on the value of \code{type}.}

\item{type, }{string, should count or density be displayed on the y-axis? If \code{"auto"}, \code{"density"} is used if \code{density} or \code{densityOverlay} is used, otherwise \code{"count"} is used. \code{"count"} preserves marginal densities if split by group, \code{"density"} re-normalizes each subgroup.}

\item{breaks, }{see \code{breaks} from \code{graphics::hist}. Additionally allows \code{"doane"} method.}

\item{histogram, }{logical, should a histogram be shown?}

\item{histogramArgs, }{list, additional arguments passed to \code{\link[ggplot2]{geom_histogram}}. Use \code{.histogramArgs} to set the options.}

\item{rug, }{logical, should rugs be shown on the x-axis?}

\item{rugArgs, }{list, additional arguments passed to \code{\link[ggplot2]{geom_rug}}. Use \code{.rugArgs} to set the options.}

\item{density, }{logical, should a density be superimposed on the plot?}

\item{densityArgs, }{logical, additional arguments passed to \code{\link[ggplot2]{geom_density}}. Use \code{.densityArgs} to set the options.}

\item{densityOverlay, }{logical, should a density overlay be superimposed on the plot?}

\item{densityOverlayArgs, }{logical, additional arguments passed to \code{\link[ggplot2]{geom_density}}. Use \code{.densityArgs} to set the options.}

\item{axisLabels, }{string, which axes should have labels displayed? If \code{"auto"}, \code{"x"} is used if \code{type == "density"}, otherwise \code{"both"} is used.}
}
\value{
\code{jaspMarginal} returns the ggplot object.
\code{.marginalArgs}, \code{histogramArgs}, \code{.rugArgs}, and \code{.densityArgs} return a list passed as arguments into their respective ggplot geoms.
}
\description{
A plot histogram with four optional components.
\describe{
\item{\code{histogram}}{Histogram which can be tweaked with \code{breaks} and \code{histogramArgs} arguments.}
\item{\code{density}}{Density line(s) which can be tweaked with \code{densityArgs}.}
\item{\code{densityOverlay}}{Density line which can be tweaked with \code{densityOverlayArgs}. Only one line is shown for the full data regardless of whether \code{group} is used.}
\item{\code{rug}}{Rugs underneath the figure which can be tweaked with \code{rugArgs}.}
}

Each of these components can be enabled (or disabled) individually.

\code{.marginalArgs}, \code{histogramArgs}, \code{.rugArgs}, and \code{.densityArgs} are helper functions for specifying the list of tweaking options for the individual components without overriding other default values.
}
\details{
Colors are taken from \code{graphOptions("palette")}.
}
\examples{
set.seed(1)
x <- rnorm(250)
jaspMarginal(x)
jaspMarginal(x, breaks = 5)
jaspMarginal(x, breaks = 'doane')

jaspMarginal(x, density = TRUE, histogram = FALSE, rug = TRUE)


group <- as.factor(sample(letters[1:2], 500, TRUE, prob = c(0.3, 0.7)))
x     <- rnorm(500, mean = c(a = 0, b = 1)[group])

jaspMarginal(x, group, density = TRUE, rug = TRUE) # does not preserve marginal proportions
jaspMarginal(x, group, density = TRUE, rug = TRUE, type = "count") # preserves marginal proportions

# stacked groups
jaspMarginal(
  x, group, density = TRUE, type = "count",
  histogramArgs = .histogramArgs(position = ggplot2::position_stack()),
  densityArgs = .densityArgs(position = ggplot2::position_stack())
  )

jaspMarginal(x, group, densityOverlay = TRUE, type = "count", histogramArgs = .histogramArgs(position = ggplot2::position_stack()))
}
