% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspMatrixPlot.R
\name{jaspMatrixPlot}
\alias{jaspMatrixPlot}
\title{Matrix plot}
\usage{
jaspMatrixPlot(
  data,
  diagonalPlotFunction = jaspHistogram,
  diagonalPlotArgs = list(),
  topRightPlotFunction = jaspBivariate,
  topRightPlotArgs = list(),
  bottomLeftPlotFunction = NULL,
  bottomLeftPlotArgs = list(),
  overwriteDiagonalAxes = "x",
  overwriteTopRightAxes = "both",
  overwriteBottomLeftAxes = "both",
  axesLabels
)
}
\arguments{
\item{data}{Data frame of data to plot.}

\item{overwriteDiagonalAxes, overwriteTopRightAxes, overwriteBottomLeftAxes}{Which axes should be overwritten such that they have a common range. Possible options:
\describe{
\item{\code{"none"}}{No axes are overwritten, hence the plots get their own scales given by \code{diagonal}, \code{topRight}, and \code{bottomLeft} functions, respectively.}
\item{\code{"both"}}{Both axes are overwritten. The plots inherit scales by setting their \code{breaks} determined by \link{getPrettyAxisBreaks}, and the plotting region is set by \link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian} with \code{limits} set to \code{range(breaks)}. Further, the name of the axis is set to \code{NULL}.}
\item{\code{"x"}}{x-axis gets overwritten (see option \code{"both"}), y-axis does not (see option \code{"none"}).}
\item{\code{"y"}}{y-axis gets overwritten (see option \code{"both"}), x-axis does not (see option \code{"none"}).}
}}

\item{axesLabels}{Optional character vector; provide column/row names of the matrix.}

\item{diagonal}{A function that draws the plots on the diagonal. Must accept arguments \code{x} (numeric), \code{xName} (character).}

\item{diagonalArgs}{A list of additional arguments to pass to \code{diagonal}.}

\item{topRight}{A function that draws the plots on the top right off-diagonal. Must accept arguments \code{x} (numeric), \code{y} (numeric), \code{xName} (character), and \code{yName} (character).}

\item{topRightArgs}{A list of additional arguments to pass to \code{topRight}.}

\item{bottomLeft}{A function that draws the plots on the bottom left off-diagonal. Must accept arguments \code{x} (numeric), \code{y} (numeric), \code{xName} (character), and \code{yName} (character).}

\item{bottomLeftArgs}{A list of additional arguments to pass to \code{bottomLeft}.}
}
\description{
Plot that consists of \code{ncol{data}} by \code{ncol{data}} plots,
where subplot on position \eqn{(i, j)} plots \code{data[, c(i, j)]}.
The plot can display three different types of plots:
\describe{
\item{\code{diagonal}}{Where \code{i == j}.}
\item{\code{topRight}}{Where \code{i < j}.}
\item{\code{bottomLeft}}{Where \code{i > j}.}
}
}
