% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspScatter.R
\name{jaspScatter}
\alias{jaspScatter}
\title{Scatter plots with optional confidence and prediction intervals.}
\usage{
jaspScatter(
  x,
  y,
  group = NULL,
  xName,
  yName,
  type = c("point", "hex", "bin", "density", "contour"),
  args = list(color = "black"),
  smooth = c("none", "lm", "glm", "gam", "loess"),
  smoothCi = FALSE,
  smoothCiLevel = 0.95,
  smoothArgs = list(),
  predict = c("none", "lm", "ellipse"),
  predictCiLevel = 0.95,
  predictArgs = list(),
  suppressAxesLabels = FALSE,
  xBreaks = NULL,
  yBreaks = NULL
)
}
\arguments{
\item{x}{Numeric vector of values on the x-axis.}

\item{y}{Numeric vector of values on the y-axis.}

\item{group}{Optional grouping variable.}

\item{xName}{Character; x-axis label. If left empty, the name of the \code{x} object is displayed. To remove the axis label, use \code{NULL}.}

\item{yName}{Character; y-axis label. If left empty, the name of the \code{y} object is displayed. To remove the axis label, use \code{NULL}.}

\item{type}{Character; How should the distribution of the data be displayed:
\describe{
\item{"point"}{Using \link{geom_point}.}
\item{"hex"}{Using \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}
\item{"contour"}{Using \link[ggplot2:geom_contour]{ggplot2::geom_contour}.}
}}

\item{smooth}{Character; passed as \code{method} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth},
unless \code{smooth == "none"}, in which case the layer si not plotted.}

\item{smoothCi}{Logical; Should confidence interval around the smooth line be plotted?
Passed as \code{se} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{smoothCiLevel}{Numeric; Confidence level of the confidence interval around the smooth line.
Passed as \code{level} argument to \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.}

\item{predict}{Character; Method for drawing the prediction interval:
\describe{
\item{"none"}{Prediction interval is not displayed.}
\item{"lm"}{Prediction interval is plotted, the confidence bands are calculated using \link[stats:predict.lm]{stats::predict.lm}.}
\item{"ellipse"}{Prediction ellipse is plotted, assuming bi-variate normal model.}
}}

\item{predictCiLevel}{Numeric; Confidence level of the prediction interval.}

\item{suppressAxesLabels}{Logical; should axis labels be suppressed.}

\item{bins, binwidth}{Arguments passed to \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}

\item{palette}{Argument passed to \link{JASPcolors}. Palette to use for drawing \link[ggplot2:geom_hex]{ggplot2::geom_hex} and \link[ggplot2:geom_contour]{ggplot2::geom_contour}.}

\item{fill}{Argument passed to \link{geom_point}.}

\item{alpha}{Argument passed to \link{geom_point} or \link[ggplot2:geom_hex]{ggplot2::geom_hex}.}

\item{smoothColor}{Color of the smooth line.}

\item{predictColor}{Color of the prediction interval.}
}
\description{
This plot consists of three layers:
\enumerate{
\item The distribution of the data displayed as \link{geom_point} or as a \link[ggplot2:geom_hex]{ggplot2::geom_hex}.
\item Smooth line through the data displayed using \link[ggplot2:geom_smooth]{ggplot2::geom_smooth}.
\item Prediction interval of y given x using \link[stats:predict.lm]{stats::predict.lm}(assuming linear relationship), or prediction ellipse assuming bivariate normal distribution.
}
}
