% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JASPScatterPlot.R
\name{JASPScatterPlot}
\alias{JASPScatterPlot}
\title{Create a scatter plot with density}
\usage{
JASPScatterPlot(x, y, group = NULL, xName = NULL, yName = NULL,
  addSmooth = TRUE, plotAbove = c("density", "histogram", "none"),
  plotRight = c("density", "histogram", "none"),
  emulateGgMarginal = FALSE, ...)
}
\arguments{
\item{x}{x variable.}

\item{y}{y variable.}

\item{group}{optional grouping variable.}

\item{xName}{name on x-axis.}

\item{yName}{name on y-axis.}

\item{addSmooth}{should a smoothed regression line be drawn?}

\item{plotAbove}{type of plot above the scatter plot.}

\item{plotRight}{type of plot right of the scatter plot.}

\item{emulateGgMarginal}{Should the result be as similar as possible to \code{\link[ggExtra]{ggMarginal}}?}

\item{...}{}
}
\description{
Create a scatter plot with density
}
\details{
The only change added when \code{emulateGgMarginal = TRUE} is that \code{ggplot2::theme(plot.margin = unit(c(0, 0, 0.25, 0.25), "cm"))}
is added to the main plot
}
\examples{
x <- rnorm(100)
y <- rgamma(100, 1, 1)

JASPScatterPlot(x, y, xName = "Normal", yName = expression(gamma))
JASPScatterPlot(x, y, xName = "Normal", yName = "Gamma", addSmooth = FALSE)

JASPScatterPlot(x, y, xName = "Normal", yName = "Gamma", plotAbove = "histogram")
JASPScatterPlot(x, y, xName = "Normal", yName = "Gamma", plotAbove = "histogram", plotRight = "histogram")

JASPScatterPlot(x, y, xName = "Normal", yName = "Gamma", plotAbove = "histogram", plotRight = "none")
JASPScatterPlot(x, y, xName = "Normal", yName = "Gamma", plotAbove = "none",      plotRight = "none")

# we can do the same with a grouping variable
data("mtcars")
debugonce(JASPScatterPlot)
JASPScatterPlot(x = mtcars$mpg, y = mtcars$disp, group = mtcars$cyl)
JASPScatterPlot(x = mtcars$mpg, y = mtcars$disp, group = mtcars$cyl, plotAbove = "histogram", plotRight = "histogram")

## NOT RUN
# spot the differences (if you find any report them to the package maintainer!)
plot <- JASPScatterPlot(x, y, xName = "Normal", yName = "Gamma", addSmooth = FALSE, emulateGgMarginal = TRUE)
plot
ggExtra::ggMarginal(plot$subplots[[1]])

}
