% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEditing.R
\name{plotEditing}
\alias{plotEditing}
\title{Edit a plot}
\usage{
plotEditing(graph, newOptions)
}
\arguments{
\item{graph}{a ggplot2 object}

\item{newOptions}{an options list}
}
\description{
Edit a plot
}
