% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPieChart.R
\name{plotPieChart}
\alias{plotPieChart}
\title{Make a Pie Chart}
\usage{
plotPieChart(value, group, legendName = deparse(substitute(group)),
  legendLabels = if (is.factor(group)) levels(group) else unique(group),
  showAxisText = TRUE, showAxisTicks = showAxisText,
  asPercentages = TRUE, ...)
}
\arguments{
\item{value}{the value for each group.}

\item{group}{character of factor that indicates which value belongs to which group.}

\item{legendName}{Character, title of the legend.}

\item{legendLabels}{character vector with names for the legend.}

\item{showAxisText}{Logical, should the axis text be shown?}

\item{asPercentages}{Logical, should value be transformed to percentages? Recommended to be true.}

\item{...}{Arguments to be passed to \code{\link{themeJasp}}.}
}
\value{
a ggplot object.
}
\description{
Make a Pie Chart
}
\examples{
value <- c(25, 25, 50)
gg <- letters[1:3]
ga <- letters[4:6]

g <- plotPieChart(value, gg)
print(g)
plotPieChart(value, gg, ga)
plotPieChart(value, gg, ga, showAxisTicks = FALSE)
plotPieChart(value, gg, ga, showAxisTicks = FALSE, legend.position = "none") # hide the legend

# axis can still be modified
print(g + scale_y_continuous(breaks = seq(0, 100, 10)))

# something more extreme:
value <- rpois(25, 10)
g <- as.character(seq_len(25))
plotPieChart(value, g)




}
