% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQQnorm.R
\name{plotQQnorm}
\alias{plotQQnorm}
\title{Make a Q-Q plot}
\usage{
plotQQnorm(
  residuals,
  lower = NULL,
  upper = NULL,
  abline = TRUE,
  ablineColor = "red",
  xName = gettext("Theoretical quantiles", domain = "R-JASPgraphs"),
  yName = gettext("Observed quantiles", domain = "R-JASPgraphs")
)
}
\arguments{
\item{residuals}{Numeric vector, the residuals of the analysis.}

\item{lower}{Numeric vector, lower confidence interval of each residual. If NULL, no error bars are drawn.}

\item{upper}{Numeric vector, lower confidence interval of each residual. If NULL, no error bars are drawn.}

\item{abline}{Logical, should an abline be drawn through the origin?}

\item{ablineColor}{String, color of the abline.}

\item{xName}{String, name for the x-axis.}

\item{yName}{String, name for the y-axis.}
}
\description{
Make a Q-Q plot
}
\details{
This function is equivalent to \code{qqnorm(residuals); qqline(residuals)}, but uses \code{ggplot2} and allows for confidence bands.
}
\examples{
x <- rnorm(100)
lower <- x - .1
upper <- x + .1
JASPgraphs::plotQQnorm(x, lower, upper)

}
