% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspScales.R
\name{scale_x_continuous}
\alias{scale_x_continuous}
\alias{scale_y_continuous}
\title{Continuous axis scales}
\usage{
scale_x_continuous(
  name = waiver(),
  breaks = getPrettyAxisBreaks,
  minor_breaks = waiver(),
  labels = axesLabeller,
  limits = "JASP",
  expand = waiver(),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  position = "bottom",
  sec.axis = waiver()
)

scale_y_continuous(
  name = waiver(),
  breaks = getPrettyAxisBreaks,
  minor_breaks = waiver(),
  labels = axesLabeller,
  limits = "JASP",
  expand = waiver(),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  position = "left",
  sec.axis = waiver()
)
}
\arguments{
\item{name}{see details}

\item{breaks}{see details}

\item{minor_breaks}{see details}

\item{labels}{see details}

\item{limits}{see details}

\item{expand}{see details}

\item{oob}{see details}

\item{na.value}{see details}

\item{trans}{see details}

\item{position}{see details}

\item{sec.axis}{see details}
}
\description{
Continuous axis scales
}
\details{
These functions are virtually identical to \code{\link[ggplot2]{scale_x_continuous}} and \code{\link[ggplot2]{scale_y_continuous}}
except that default values are different, these use a different function to determine the default
axis breaks.
}
