% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeJasp.R
\name{themeJasp}
\alias{themeJasp}
\alias{themeJaspRaw}
\title{JASP theme for ggplot2 objects}
\usage{
themeJasp(
  graph,
  xAxis = TRUE,
  yAxis = TRUE,
  sides = "bl",
  axis.title.cex = getGraphOption("axis.title.cex"),
  bty = getGraphOption("bty"),
  fontsize = getGraphOption("fontsize"),
  family = getGraphOption("family"),
  horizontal = FALSE,
  legend.position = getGraphOption("legend.position"),
  legend.justification = "top",
  axisTickLength = getGraphOption("axisTickLength"),
  axisTickWidth = getGraphOption("axisTickWidth")
)

themeJaspRaw(
  legend.position = "none",
  legend.cex = 1,
  axis.title.cex = 1,
  family = getGraphOption("family"),
  axisTickLength = getGraphOption("axisTickLength"),
  axisTickWidth = getGraphOption("axisTickWidth"),
  fontsize = getGraphOption("fontsize"),
  legend.justification = "top",
  legend.title = element_text(family = family, size = fontsize, hjust = 0.5),
  Xvjust = NULL,
  Yvjust = NULL
)
}
\arguments{
\item{graph}{a ggplot2 object}

\item{xAxis}{Should there be a horizontal line for the x-axis?}

\item{yAxis}{Should there be a horizontal line for the x-axis?}

\item{sides}{see \code{\link{geom_rangeframe}}}

\item{axis.title.cex}{scalar magnification for the title of the axes.}

\item{bty}{remake R's bty  = 'n'?}

\item{fontsize}{global font size.}

\item{family}{global font familiy.}

\item{horizontal}{flip graph?}

\item{legend.position}{where should the legend be?}

\item{legend.justification}{see theme}

\item{axisTickLength}{length of axis ticks.}

\item{axisTickWidth}{width of axis ticks.}

\item{legend.cex}{magnification of the font size for the legend.}

\item{legend.title}{should legend have a title?}

\item{Xvjust}{adjustment of x-axis tick labels}

\item{Yvjust}{adjustment of y-axis tick labels}
}
\description{
Set a JASP theme and additional options with \code{themeJASP(graph)}, or add the theme to a graph like
with other themes, \code{graph + themeJaspRaw()}.
}
