% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeJasp.R
\name{themeJasp}
\alias{themeJasp}
\title{Deprecated: use \link{themeJaspRaw} and \link{geom_rangeframe} instead.}
\usage{
themeJasp(
  graph,
  xAxis = TRUE,
  yAxis = TRUE,
  sides = "bl",
  axis.title.cex = getGraphOption("axis.title.cex"),
  bty = getGraphOption("bty"),
  fontsize = getGraphOption("fontsize"),
  family = getGraphOption("family"),
  horizontal = FALSE,
  legend.position = getGraphOption("legend.position"),
  legend.justification = "top",
  axisTickLength = getGraphOption("axisTickLength"),
  axisTickWidth = getGraphOption("axisTickWidth")
)
}
\arguments{
\item{graph}{a ggplot2 object}

\item{xAxis}{Should there be a horizontal line for the x-axis?}

\item{yAxis}{Should there be a horizontal line for the x-axis?}

\item{sides}{see \link{geom_rangeframe}}

\item{axis.title.cex}{scalar magnification for the title of the axes.}

\item{bty}{remake R's bty  = 'n'?}

\item{fontsize}{global font size.}

\item{family}{global font familiy.}

\item{horizontal}{flip graph?}

\item{legend.position}{where should the legend be?}

\item{legend.justification}{see theme}

\item{axisTickLength}{length of axis ticks.}

\item{axisTickWidth}{width of axis ticks.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link{themeJaspRaw}} and \link{geom_rangeframe} instead.
}
